Change the get current place details API code form MapPlace Seach Activity to Landing Page MapView Fragment. Change the current fetch location Button form the source text view and written it programmatically to fetch current location and get the current details via google map API Bug fix on getting current location details from google API  if the location is not fetched.
Changed all the phone call numbers to dynamic.
discussion with George on the delay of the current location time delay fetching location
System crash fixes with the help of Nidhessh, USB ports got the disabled.
Bug fix on the API side, to help Meghna
Had a discussion with Meghna to append the API location on both Driver Accept trip and user side Confirm trip
Worked on the newly added Push notification variable to get Drop location and Append it to draw the polynomial maps.